requirements.txt
# FastAPI and ASGI server
fastapi==0.95.1
uvicorn==0.21.1

# HTTP clients for synchronous/asynchronous API calls
httpx==0.24.1
aiofiles==23.1.0

# Environment variable management
python-dotenv==1.0.0

# Data validation and settings management
pydantic==1.10.7

# Embedding and ML (choose one or both as needed)
sentence-transformers==2.2.2
openai==0.27.0

# Caching (if using Redis)
redis==4.5.4
aioredis==2.0.1

# Testing tools
pytest==7.2.2
pytest-asyncio==0.20.2